+++ application/app/Http/Controllers/FundsController.php
@@ -168,6 +168,7 @@ class FundsController extends Controller
             'amountRemaining' => $amountRemaining,
             'tallies' => $this->getTallies($activeFund, $perPage, $page),
             'filters' => $this->queryParams,
+            'quickPitches' => $this->getActiveFundQuickPitches($activeFund, $proposals),
         ]);
     }
 
@@ -210,6 +211,42 @@ class FundsController extends Controller
         }
 
         return $query->get();
+    }
+
+    private function getActiveFundQuickPitches(Fund $fund, ProposalRepository $proposals)
+    {
+        try {
+            $rawProposals = $proposals
+                ->with(['users', 'campaign', 'fund'])
+                ->whereNotNull('quickpitch')
+                ->where('fund_id', $fund->id)
+                ->limit(15)
+                ->get();
+
+            if ($rawProposals->count() < 3) {
+                return [
+                    'featured' => collect([]),
+                    'regular' => ProposalData::collect($rawProposals),
+                ];
+            }
+
+            $featuredRaw = $rawProposals->random(3);
+
+            $featuredIds = $featuredRaw->pluck('id');
+            $regularRaw = $rawProposals->whereNotIn('id', $featuredIds);
+
+            return [
+                'featured' => ProposalData::collect($featuredRaw),
+                'regular' => ProposalData::collect($regularRaw),
+            ];
+        } catch (\Throwable $e) {
+            report($e);
+
+            return [
+                'featured' => collect([]),
+                'regular' => collect([]),
+            ];
+        }
     }
 
     private function getProposals(Fund $activeFund, ProposalRepository $proposals)
@@ -416,7 +453,7 @@ class FundsController extends Controller
 
                 $fundRanking = null;
                 if ($tally->metas && $tally->metas->isNotEmpty()) {
-                    $fundRankMeta = $tally->metas->firstWhere('key', 'overall_rank');
+                    $fundRankMeta = $tally->metas->firstWhere('key', 'fund_rank');
                     if ($fundRankMeta) {
                         $fundRanking = (int) $fundRankMeta->content;
                     }